requirements.txt
streamlit
langchain-community
sentence-transformers
faiss-cpu
pypdf
tiktoken
google-generativeai
